Remove height limitations on #root
This bit of CSS causes the navbar to be smashed in the vertical direction on pages that have a lot of stuff in them (like, say, on a page with a well-populated list of events).

While this sort of thing might be useful later on if we get a sticky footer, I think it’s best to remove it instead of band-aiding the navbar with a `flex-shrink: 0` constraint.